Rename: Fix references broken by move/rename
